+++ yorm/decorators.py
@@ -37,7 +37,7 @@ def sync_object(instance, path, attrs=None, **kwargs):
 
     :param auto_create: automatically create the file to save attributes
     :param auto_save: automatically save attribute changes to the file
-    :param auto_attr: automatically add new attributes from the file
+    :param auto_track: automatically add new attributes from the file
 
     """
     log.info("Mapping %r to %s...", instance, path)
@@ -46,18 +46,17 @@ def sync_object(instance, path, attrs=None, **kwargs):
     patch_methods(instance)
 
     attrs = _ordered(attrs) or common.attrs[instance.__class__]
-    kwargs['auto_store'] = kwargs.pop('auto_save', True)
     mapper = Mapper(instance, path, attrs, **kwargs)
 
     if mapper.missing:
         if mapper.auto_create:
             mapper.create()
-            if mapper.auto_store:
-                mapper.store()
-                mapper.fetch()
+            if mapper.auto_save:
+                mapper.save()
+                mapper.load()
     else:
-        if mapper.auto_store:
-            mapper.fetch()
+        if mapper.auto_save:
+            mapper.load()
 
     common.set_mapper(instance, mapper)
     log.info("Mapped %r to %s", instance, path)
@@ -74,7 +73,7 @@ def sync_instances(path_format, format_spec=None, attrs=None, **kwargs):
 
     :param auto_create: automatically create the file to save attributes
     :param auto_save: automatically save attribute changes to the file
-    :param auto_attr: automatically add new attributes from the file
+    :param auto_track: automatically add new attributes from the file
 
     """
     format_spec = format_spec or {}